scores added
